Fix test by extracting arg parsing to separate cli module that doesn't need to access seqrepo
