+++ test/test_cli.py
@@ -1,8 +1,9 @@
-from clinvar_gk_pilot.main import parse_args
+from clinvar_gk_pilot.cli import parse_args
 
 
 def test_parse_args():
     argv = ["--filename", "test.txt"]
     opts = parse_args(argv)
     assert opts["filename"] == "test.txt"
-    assert len(opts) == 1
+    assert opts["parallelism"] == 1
+    assert len(opts) == 2
